# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Anime Calendar API
  version: 1.0.0
  description: |
    REST API for Anime Calendar — create and manage personal anime/manga calendars
    sourced from AniList, exportable as iCalendar (.ics) feeds.

    ## Authentication
    Most endpoints require a Bearer JWT obtained via `/login`, `/register`, or `/auth/google`.
    Include it as: `Authorization: Bearer <token>`

    ## Error responses
    All errors return a JSON body: `{"error": "<message>"}`.
    The HTTP status code indicates the error category.
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # -------------------------------------------------------------------------
  # Reusable response components
  # -------------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
...
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Invalid request data
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Requested resource not found
...
    GoogleOAuthRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Google ID token obtained from the Google Identity Services JS library
          example: eyJhbGciOiJSUzI1NiIsImtpZCI6...
    GoogleOAuthResponse:
      type: object
      required: [token, username, email, avatar]
      properties:
        token:
          type: string
          description: JWT bearer token (24h expiry)
        username:
          type: string
          description: User's display name from Google profile
        email:
          type: string
          format: email
        avatar:
          type: string
          format: uri
          description: Google profile picture URL
...
    Language:
      type: string
      enum: [english, native, romaji]
      description: Title language preference for AniList titles
      default: english
    Item:
      type: object
      required: [id, id_mal, title, media_type, episode_duration, airing_schedule, cover_image, banner_image]
      properties:
        id:
          type: integer
          description: AniList media ID
          example: 1535
        id_mal:
          type: integer
          nullable: true
          description: MyAnimeList ID (nullable)
          example: 16498
        title:
          type: object
          properties:
            english:
              type: string
              nullable: true
              example: Attack on Titan
            native:
              type: string
              nullable: true
              example: 進撃の巨人
            romaji:
              type: string
              nullable: true
              example: Shingeki no Kyojin
        media_type:
          type: string
          enum: [ANIME, MANGA]
        episode_duration:
          type: integer
          description: Episode duration in minutes
          example: 24
        airing_schedule:
          type: array
          items:
            $ref: '#/components/schemas/AiringSchedule'
        cover_image:
          type: object
          properties:
            medium:
              type: string
              format: uri
              nullable: true
            large:
              type: string
              format: uri
              nullable: true
            extra_large:
              type: string
              format: uri
              nullable: true
            color:
              type: string
              nullable: true
              example: '#2b2d42'
        banner_image:
          type: string
          format: uri
          nullable: true
        recommendations:
          type: array
          items:
            $ref: '#/components/schemas/Recommendation'
    AiringSchedule:
      type: object
      required: [episode, airing_at]
      properties:
        episode:
          type: integer
          example: 1
        airing_at:
          type: integer
          description: Unix timestamp of scheduled air time
          example: 1680000000
    Recommendation:
      type: object
      required: [rating, media]
      properties:
        rating:
          type: integer
          description: Community rating for this recommendation
          example: 42
        media:
          type: object
          required: [id, id_mal, title, cover_image]
          properties:
            id:
              type: integer
            id_mal:
              type: integer
              nullable: true
            title:
              type: object
              properties:
                english:
                  type: string
                  nullable: true
                native:
                  type: string
                  nullable: true
                romaji:
                  type: string
                  nullable: true
            cover_image:
              type: object
              properties:
                medium:
                  type: string
                  format: uri
                  nullable: true
                large:
                  type: string
                  format: uri
                  nullable: true
                extra_large:
                  type: string
                  format: uri
                  nullable: true
                color:
                  type: string
                  nullable: true
...
    UserResponse:
      type: object
      required: [username, email, is_oauth]
      properties:
        username:
          type: string
          description: Empty string for OAuth-only accounts
          example: animefan42
        email:
          type: string
          format: email
          example: user@example.com
        is_oauth:
          type: boolean
          description: True if the account was created via Google OAuth (no local password)
          example: false
    UpdateUserRequest:
      type: object
      required: [username, email]
      properties:
        username:
          type: string
          maxLength: 50
          example: newname
        email:
          type: string
          format: email
          example: newemail@example.com
    UpdatePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          format: password
        new_password:
          type: string
          format: password
          maxLength: 128
    UserSettings:
      type: object
      required: [theme_preference, language_preference, title_language_preference, timezone]
      properties:
        user_id:
          type: integer
          readOnly: true
        theme_preference:
          type: string
          enum: [light, dark]
          default: dark
        language_preference:
          type: string
          enum: [en, pt]
          default: en
          description: UI language
        title_language_preference:
          $ref: '#/components/schemas/Language'
        timezone:
          type: string
          description: IANA timezone name
          example: America/Sao_Paulo
          default: UTC
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true

# ---------------------------------------------------------------------------
# Default security (overridden where not required)
# ---------------------------------------------------------------------------
security:
  - bearerAuth: []

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /auth/google:
    post:
      tags: [Auth]
      summary: Sign in with Google (ID token flow)
      operationId: googleOAuth
      description: |
        Validates a Google ID token (obtained from the Google Identity Services JS library).
        Creates the user account on first login. Returns a JWT for subsequent requests.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleOAuthRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoogleOAuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /items/search:
    get:
      tags: [Items]
      summary: Search AniList for anime/manga by name
      operationId: searchItems
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
          example: Attack on Titan
        - name: media_type
          in: query
          required: false
          schema:
            type: string
            enum: [ANIME, MANGA, '']
          description: Filter by media type; omit for all types
      responses:
        '200':
          description: Matching items from AniList
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /items:
    get:
      tags: [Items]
      summary: Fetch multiple items from AniList by IDs
      operationId: getItems
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: array
            items:
              type: integer
          style: form
          explode: true
          description: List of AniList media IDs
      responses:
        '200':
          description: Item data for each requested ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ----------- User ---------------------------------------------------------
  /user/details:
    get:
      tags: [User]
      summary: Get profile details for the current user
      operationId: getUserDetails
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user:
    put:
      tags: [User]
      summary: Update username and email
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [User]
      summary: Delete the current user account
      operationId: deleteUser
      responses:
        '204':
          description: Account deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/password:
    put:
      tags: [User]
      summary: Change the account password
      operationId: updatePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          description: Password updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/settings:
    get:
      tags: [User]
      summary: Get user preferences
      operationId: getUserSettings
      responses:
        '200':
          description: User settings (returns defaults if never saved)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      tags: [User]
      summary: Update user preferences
      operationId: updateUserSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserSettings'
      responses:
        '200':
          description: Settings saved
        '401':
          $ref: '#/components/responses/Unauthorized'